# artifacts/api-server/src/main/java/com/flowforge/dto/auth/AuthResponse.java
package com.flowforge.dto.auth;

import com.flowforge.dto.user.UserResponse;
import lombok.AllArgsConstructor;
import lombok.Data;

@Data
@AllArgsConstructor
public class AuthResponse {
    private String token;
    private UserResponse user;
}