# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Multi-Tenant Workflow Automation Engine API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user and create/join a tenant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized

  # ── Tenants ──────────────────────────────────────────────────────────────────
...
  /tenants/{tenantId}:
    get:
      operationId: getTenant
      tags: [tenants]
      summary: Get a specific tenant
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Tenant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
        "404":
          description: Not found
    patch:
      operationId: updateTenant
      tags: [tenants]
      summary: Update tenant details (ORG_ADMIN)
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TenantUpdate"
      responses:
        "200":
          description: Updated tenant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"

  # ── Users ────────────────────────────────────────────────────────────────────
...
  /executions:
    get:
      operationId: listExecutions
      tags: [executions]
      summary: List execution logs for the tenant
      parameters:
        - name: workflowId
          in: query
          schema:
            type: ["integer", "null"]
        - name: status
          in: query
          schema:
            type: string
            enum: [success, failure, running, all]
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Execution logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Execution"
  /executions/{executionId}:
    get:
      operationId: getExecution
      tags: [executions]
      summary: Get a specific execution log
      parameters:
        - name: executionId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Execution log
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Execution"
        "404":
          description: Not found
  /executions/{executionId}/retry:
    post:
      operationId: retryExecution
      tags: [executions]
      summary: Retry a failed execution
      parameters:
        - name: executionId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "202":
          description: Retry started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Execution"

  # ── Dashboard ────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    RegisterInput:
      type: object
      required: [name, email, password, organizationName]
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        organizationName:
          type: string
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"

    # ── Tenant ──────────────────────────────────────────────────────────────────
    Tenant:
      type: object
      required: [id, name, slug, plan, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        plan:
          type: string
          enum: [free, pro, enterprise]
        userCount:
          type: ["integer", "null"]
        workflowCount:
          type: ["integer", "null"]
        createdAt:
          type: string
    TenantUpdate:
      type: object
      properties:
        name:
          type: string
        plan:
          type: string
          enum: [free, pro, enterprise]

    # ── User ────────────────────────────────────────────────────────────────────
    User:
      type: object
      required: [id, name, email, role, tenantId, active, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [SUPER_ADMIN, ORG_ADMIN, USER]
        tenantId:
          type: integer
        active:
          type: boolean
        createdAt:
          type: string
...
    Execution:
      type: object
      required: [id, workflowId, tenantId, status, startedAt]
      properties:
        id:
          type: integer
        workflowId:
          type: integer
        workflowName:
          type: ["string", "null"]
        tenantId:
          type: integer
        status:
          type: string
          enum: [success, failure, running]
        logs:
          type: ["string", "null"]
        errorMessage:
          type: ["string", "null"]
        durationMs:
          type: ["integer", "null"]
        startedAt:
          type: string
        completedAt:
          type: ["string", "null"]

    # ── Dashboard ───────────────────────────────────────────────────────────────
...